RUST-PLI-PREPROCESSOR/pli_tokenizer/src/modules/macro_expander.rs
#![allow(unused_imports)] // Suppress unused warnings for imports temporarily.
                          ////////////////////////////////////////////////////////////////////////////////
                          // MODULE NAME: Macro Expander
                          // -----------------------------------------------------------------------------
                          // Description:
                          // This module is responsible for expanding macros within PL/I preprocessor
                          // files. A macro is a reusable code block or substitution directive defined
                          // within the source file. The macro expander parses and expands such definitions.
                          //
                          // Features:
                          // - Parses macro definitions from PL/I preprocessor directives.
                          // - Expands macros based on input parameters and definitions.
                          // - Handles nested and recursive macro calls.
                          // - Supports validation and error checking for undefined or malformed macros.
                          //
                          // Purpose:
                          // The macro expander simplifies repetitive code and enhances modularity within
                          // PL/I preprocessor files by substituting macros with their expanded content.
                          //
                          // Usage:
                          // 1. Use `expand_macro` to expand a macro definition or usage.
                          // 2. Integrate with the tokenizer to handle macros inline.
                          //
                          // Example:
                          // ```rust
                          // use macro_expander::expand_macro;
                          //
                          // let input = "%MACRO TEST; VALUE = 1; %ENDMACRO;";
                          // let result = expand_macro(input);
                          // assert_eq!(result, Some(expanded_output));
                          // ```
                          //
                          // Dependencies:
                          // - `log`: For logging during macro processing.
                          // - `regex`: For parsing macro patterns (to be implemented).
                          //
                          // Notes:
                          // - This module is currently a skeleton and will be implemented in future iterations.
                          // - Placeholder functions and structures are provided for modular development.
                          //
                          // Enhancements:
                          // - Add support for parameterized macros.
                          // - Integrate with tokenizer for seamless expansion during tokenization.
                          //
                          // Author: Jean-Pierre Sainfeld
                          // Assistant: ChatGPT
                          // Company: FirstLink Consulting Services (FLCS)
                          // -----------------------------------------------------------------------------
                          ////////////////////////////////////////////////////////////////////////////////

use log::{debug, error, info, warn}; // For logging macro expansion process.
use regex::Regex; // For future implementation of macro parsing (not yet in use).

/// Expands a macro definition or usage within a given PL/I line or block of code.
///
/// # Arguments
/// - `input`: A `&str` representing the PL/I code that may contain macros.
///
/// # Returns
/// - `Option<String>`: The expanded code if macro expansion is successful,
///   or `None` if no macro expansion was performed.
///
/// # Example
/// ```rust
/// let input = "%MACRO TEST; VALUE = 1; %ENDMACRO;";
/// let result = expand_macro(input);
/// assert_eq!(result, Some("Expanded macro output"));
/// ```
///
/// # Notes
/// - This is a placeholder function for future implementation.
/// - Currently, it logs the input and returns `None`.
pub fn expand_macro(input: &str) -> Option<String> {
    // Placeholder: Log the macro expansion attempt.
    debug!(
        "expand_macro: Attempting to expand macro in input: {}",
        input
    );

    // TODO: Implement macro parsing and expansion logic here.
    warn!("expand_macro: Macro expansion logic not yet implemented.");

    None // Return None as macro expansion is not yet implemented.
}

/// Validates a macro definition for correctness (to be implemented).
///
/// # Arguments
/// - `macro_definition`: A `&str` containing the macro definition to validate.
///
/// # Returns
/// - `bool`: `true` if the macro definition is valid, otherwise `false`.
///
/// # Example
/// ```rust
/// let macro_def = "%MACRO TEST; VALUE = 1; %ENDMACRO;";
/// assert!(validate_macro(macro_def));
/// ```
pub fn validate_macro(macro_definition: &str) -> bool {
    // Placeholder: Log the validation attempt.
    debug!(
        "validate_macro: Validating macro definition: {}",
        macro_definition
    );

    // TODO: Implement macro validation logic here.
    warn!("validate_macro: Macro validation logic not yet implemented.");

    false // Return false as validation logic is not yet implemented.
}
