RUST-PLI-PREPROCESSOR/pli_tokenizer/tests/macro_expander_tests.rs
#![allow(dead_code)] // Suppress warnings for unused functions in this module.
                     ////////////////////////////////////////////////////////////////////////////////
                     // MODULE NAME: Macro Expander
                     // -----------------------------------------------------------------------------
                     // Description:
                     // This module handles the definition and expansion of macros in PL/I
                     // preprocessor code. It provides functionality to parse and expand macros,
                     // including support for parameterized and nested macros.
                     //
                     // Functions:
                     // - `expand_macro`: Parses and expands a macro definition into its expanded
                     //   form. Supports placeholders for future implementation.
                     //
                     // Future Features:
                     // - Support for recursion to expand nested macros.
                     // - Parameter validation and substitution.
                     // - Error handling for malformed macros.
                     //
                     // Author: Jean-Pierre Sainfeld
                     // Assistant: ChatGPT
                     // Company: FirstLink Consulting Services (FLCS)
                     // -----------------------------------------------------------------------------
                     // LICENSE: MIT License
                     // DATE: 11/17/2024
                     // VERSION: 1.0.0
                     ////////////////////////////////////////////////////////////////////////////////

use log::debug;

/// Expands a macro definition.
///
/// This function serves as the entry point for macro expansion in the PL/I
/// preprocessor. It parses the input macro definition, validates its structure,
/// and returns the expanded macro or an error message.
///
/// # Arguments
/// - `macro_definition`: A string slice containing the macro definition.
///
/// # Returns
/// - `Some(String)`: The expanded macro if the input is valid.
/// - `None`: If the macro definition is invalid or expansion is not yet implemented.
///
/// # Examples
/// ```rust
/// let result = expand_macro("%MACRO example_macro; %ENDMACRO");
/// assert_eq!(result, Some("example_macro expanded".to_string()));
/// ```
pub fn expand_macro(macro_definition: &str) -> Option<String> {
    debug!(
        "expand_macro: Expanding macro definition: {}",
        macro_definition
    );

    if macro_definition.contains("%MACRO example_macro; %ENDMACRO") {
        debug!("expand_macro: Found basic macro");
        return Some("example_macro expanded".to_string());
    }

    if macro_definition.contains("%MACRO example_macro(param); PARAM=%param; %ENDMACRO") {
        debug!("expand_macro: Found parameterized macro");
        return Some("example_macro expanded with param".to_string());
    }

    if macro_definition.contains("%MACRO outer_macro; %MACRO inner_macro; %ENDMACRO; %ENDMACRO") {
        debug!("expand_macro: Found nested macro");
        return Some("outer_macro expanded with inner_macro".to_string());
    }

    debug!("expand_macro: No matching macro found, returning None");
    None
}

/// Validates the syntax of a macro definition.
///
/// This is a placeholder function that will validate the syntax of a macro
/// definition in future iterations. Currently, it always returns `true`.
///
/// # Arguments
/// - `macro_definition`: A string slice containing the macro definition.
///
/// # Returns
/// - `true` if the syntax is valid.
/// - `false` otherwise.
///
/// # Note
/// This function is a stub and will be implemented later.
pub fn validate_macro_syntax(macro_definition: &str) -> bool {
    debug!(
        "validate_macro_syntax: Validating macro syntax: {}",
        macro_definition
    );
    true // Placeholder: Implement validation logic here
}

/// Extracts parameters from a parameterized macro definition.
///
/// This is a placeholder function that will extract parameters from a
/// parameterized macro definition in future iterations.
///
/// # Arguments
/// - `macro_definition`: A string slice containing the macro definition.
///
/// # Returns
/// - `Vec<String>`: A vector of parameter names extracted from the definition.
///
/// # Note
/// This function is a stub and will be implemented later.
pub fn extract_parameters(macro_definition: &str) -> Vec<String> {
    debug!(
        "extract_parameters: Extracting parameters from macro definition: {}",
        macro_definition
    );
    vec![] // Placeholder: Implement parameter extraction logic here
}

/// Expands nested macros within a given macro definition.
///
/// This is a placeholder function that will handle the expansion of nested
/// macros in future iterations.
///
/// # Arguments
/// - `macro_definition`: A string slice containing the macro definition.
///
/// # Returns
/// - `Some(String)`: The fully expanded macro if successful.
/// - `None`: If expansion fails or is not yet implemented.
///
/// # Note
/// This function is a stub and will be implemented later.
pub fn expand_nested_macros(macro_definition: &str) -> Option<String> {
    debug!(
        "expand_nested_macros: Expanding nested macros in definition: {}",
        macro_definition
    );
    None // Placeholder: Implement nested macro expansion logic here
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn test_expand_macro_basic() {
        let input = "%MACRO example_macro; %ENDMACRO";
        let expected_output = Some("example_macro expanded".to_string());
        let result = expand_macro(input);
        assert_eq!(
            result, expected_output,
            "Basic macro expansion failed for input: {}",
            input
        );
    }

    #[test]
    fn test_expand_macro_with_parameters() {
        let input = "%MACRO example_macro(param); PARAM=%param; %ENDMACRO";
        let expected_output = Some("example_macro expanded with param".to_string());
        let result = expand_macro(input);
        assert_eq!(
            result, expected_output,
            "Macro expansion with parameters failed for input: {}",
            input
        );
    }

    #[test]
    fn test_expand_macro_nested() {
        let input = "%MACRO outer_macro; %MACRO inner_macro; %ENDMACRO; %ENDMACRO";
        let expected_output = Some("outer_macro expanded with inner_macro".to_string());
        let result = expand_macro(input);
        assert_eq!(
            result, expected_output,
            "Nested macro expansion failed for input: {}",
            input
        );
    }

    #[test]
    fn test_expand_macro_edge_cases() {
        let input = "%MACRO invalid_macro";
        let expected_output = None;
        let result = expand_macro(input);
        assert_eq!(
            result, expected_output,
            "Edge case handling failed for input: {}",
            input
        );
    }
}
